move wsgi base to private utils module and add test for base class
